Restructured directory to make more sense
