+++ src/main.rs
@@ -1,5 +1,6 @@
-mod application;
-mod signal_handler;
+pub mod application;
+pub mod config;
+pub mod date_checker;
 
 use application::AppError;
